+++ src/applications/content-playlists-app/playlists-utils.service.ts
@@ -1,5 +1,5 @@
-import { Injectable } from '@angular/core';
-import { KalturaPlaylist } from 'kaltura-ngx-client';
+import {Injectable} from '@angular/core';
+import { KalturaPlaylist, KalturaPlaylistType } from 'kaltura-ngx-client';
 
 @Injectable()
 export class PlaylistsUtilsService {
@@ -9,5 +9,8 @@ export class PlaylistsUtilsService {
 
   public isRapt(playlist: KalturaPlaylist): boolean {
       return playlist.adminTags && playlist.adminTags.split(',').indexOf('raptentry') > -1 ? true : false;
+  }
+  public isPath(playlist: KalturaPlaylist): boolean {
+      return playlist.playlistType === KalturaPlaylistType.path;
   }
 }